PSG-1171: move concat_csv.py to scripts/common/
